+++ src/snowflake/snowpark/mock/snowflake_data_type.py
@@ -118,7 +118,7 @@ class ColumnEmulator(pd.Series):
     def __add__(self, other):
         result = super().__add__(other)
         # TODO: set sf_type for result by converting from pandas dtype to sf type?
-        # result.sf_type = calculate_type(self.sf_type, other.sf_type, op="+")
+        result.sf_type = calculate_type(self.sf_type, other.sf_type, op="+")
         return result
 
     def __radd__(self, other):
