+++ tests/integ/scala/test_dataframe_suite.py
@@ -179,7 +179,7 @@ def test_adjust_column_width_of_show(session):
 
 
 @pytest.mark.localtest
-def oh_null_data(session):
+def test_show_with_null_data(session):
     df = session.create_dataframe([[1, None], [2, "NotNull"]]).to_df("a", "b")
     # run show(), make sure no error is reported
     df.show(10)
@@ -240,7 +240,7 @@ def test_show(session):
     # make sure show runs with sql
     session.sql("show tables").show()
 
-    session.sql("drop table if exists Dable_123").show()
+    session.sql("drop table if exists test_table_123").show()
 
     # truncate result, no more than 50 characters
     res = session.sql("drop table if exists test_table_123")._show_string(1)
